fix(v0.36.33): institutional_holdings.total_shares BIGINT (v0.36.25 follow-up)
v0.36.25's URL fix made the institutional_ownership collector return real
data, which then failed nightly with "integer out of range". total_shares
was INTEGER (PG int32, max 2.1B); aggregate institutional share counts for
megacaps exceed it — AAPL aggregates to 9.9B across 8184 holders.

SQLite never hit this (int64 affinity); strict PG integer overflowed.

Fix:
- registry total_shares INTEGER -> BIGINT
- postgres.py _TYPE_MAP: add BIGINT -> BIGINT (first BIGINT column)
- PG migrated: ALTER COLUMN total_shares TYPE bigint
- verified: collect_institutional_ownership('AAPL') stores 9893657756 OK

No restart needed — collector sends unbounded Python int, PG column now
bigint; in-memory registry version irrelevant to the INSERT. Also avoids
a restart in the 21:30-22:30 ET overnight window.

Tests: NEW tests/test_institutional_holdings_bigint_v0_36_33.py (3) + schema
suite = 51 pass.

+++ src/schema/registry.py
@@ -2679,7 +2679,12 @@ _register(TableDef(
         ColumnDef("ticker", "TEXT", nullable=False),
         ColumnDef("as_of_date", "TEXT", nullable=False,
                   description="YYYY-MM-DD — quarterly filing date"),
-        ColumnDef("total_shares", "INTEGER", nullable=True),
+        # BIGINT (v0.36.33): aggregate institutional share count across all
+        # holders of a megacap exceeds int32's 2.1B ceiling (AAPL ~8000
+        # holders; collector sums `share` per holder). PG int32 overflowed
+        # with "integer out of range" once v0.36.25's URL fix started
+        # returning real data.
+        ColumnDef("total_shares", "BIGINT", nullable=True),
         ColumnDef("num_holders", "INTEGER", nullable=True),
         ColumnDef("top_5_holders_pct", "REAL", nullable=True),
         ColumnDef("qoq_delta_pct", "REAL", nullable=True,
